+++ go/models/response_types.go
@@ -113,6 +113,17 @@ func CreateNilKeyWithArrayOfMembersAndScoresResult() Result[KeyWithArrayOfMember
 	return Result[KeyWithArrayOfMembersAndScores]{val: KeyWithArrayOfMembersAndScores{"", nil}, isNil: true}
 }
 
+func CreateRankAndScoreResult(
+	rank int64,
+	score float64,
+) Result[RankAndScore] {
+	return Result[RankAndScore]{val: RankAndScore{Rank: rank, Score: score}, isNil: false}
+}
+
+func CreateNilRankAndScoreResult() Result[RankAndScore] {
+	return Result[RankAndScore]{val: RankAndScore{Rank: 0, Score: 0}, isNil: true}
+}
+
 // Enum to distinguish value types stored in `ClusterValue`
 type ValueType int
 
@@ -435,4 +446,12 @@ type KeyValues struct {
 	Key string
 	// The slice of string values associated with the key
 	Values []string
+}
+
+// RankAndScore represents the rank and score of a given member
+type RankAndScore struct {
+	// The rank of the member
+	Rank int64
+	// The score of the member
+	Score float64
 }